-1- start.sh

+++ docker/server/start.sh
@@ -1,6 +1,6 @@
 #! /bin/bash
 
-if [ ! -f /var/www/.env ] [ ! -f /var/www/vendor/autoload.php ]; then
+if [ ! -f /var/www/.env ] && [ ! -f /var/www/vendor/autoload.php ]; then
     cd /var/www
     cp .env.example .env
     php artisan key:generate
